Create BankersRoulettee.py
Program to randomly choose a payer for the bill amongst a group of people. 
The term banker's roulette is a play on Russian Roulette, a dangerous game.  It is coined due to a culture where rich bankers would go on spending sprees usually in fancy restaurants and play a game to randomly select who pays the bill.