Update BankersRoulette.py
Program to randomly choose a payer for the bill amongst a group of people.
Banker's Roulette is a wordplay on Russian Roulette, a dangerous game. It origins from a culture where wealthy bankers would go on spending sprees, usually in fancy restaurants, and randomly decide on who pays the bill through a game.