Create FizzBuzz.py
FizzBuzz is a game where a group of people take turns reciting numbers starting from 1 to any range, with the catch being every time a number divisible by 3 occurs, the player must say "FIZZ". For numbers divisible by 5, the word is "BUZZ" and "FIZZBUZZ" for numbers divisible by 3 and 5.
The program generates the "answers" to the game.