Create ChoiceHelper.py
A program to help you choose a subject to study when you need clarification about where to start.
The program requires you to rate each subject across 5 factors and gives you the optimal subject(s) to work on based on the ratings.

This is a personal project to solve an inconvenience that frequents my studying journey (and hopefully yours too). This is NOT a part of the Udemy bootcamp.